requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
cycler==0.10.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mysql-connector-python==8.0.16
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==5.0.0
protobuf==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.0.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
